# src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  description: |
    This is the OpenRQM API specification.
    [OpenRQM Docs - Github](https://github.com/openrqm/openrqm-docs)
  license:
    name: LGPL-2.0-only
    url: https://spdx.org/licenses/LGPL-2.0-only.html#licenseText
  title: OpenRQM
  version: 1.0.1
...
servers:
- url: https://virtserver.swaggerhub.com/OpenRQM/OpenRQM/1.0.0
security:
- bearerAuth: []
...
paths:
...
  /document:
    delete:
      operationId: deleteDocument
      parameters:
      - description: The document to delete
        explode: true
        in: query
        name: documentId
        required: true
        schema:
          format: int64
          type: integer
        style: form
      responses:
        "200":
          description: The document was deleted successfully
      summary: Delete a document
      tags:
      - Documents
      x-accepts: application/json
      x-tags:
      - tag: Documents
    get:
      operationId: getDocument
      parameters:
      - description: The document identifier
        explode: true
        in: query
        name: documentId
        required: true
        schema:
          format: int64
          type: integer
        style: form
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RQMDocument'
          description: Document successfully fetched
      summary: Get a document
      tags:
      - Documents
      x-accepts: '*/*'
      x-tags:
      - tag: Documents
    patch:
      operationId: patchDocument
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RQMDocument'
        description: The document to update
        required: true
      responses:
        "200":
          description: Updating document successful
      summary: Update a document
      tags:
      - Documents
      x-contentType: application/json
      x-accepts: application/json
      x-tags:
      - tag: Documents
    post:
      operationId: postDocument
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RQMDocument'
        description: The document to create
        required: true
      responses:
        "200":
          description: The document created by the backend
      summary: Add a document
      tags:
      - Documents
      x-contentType: application/json
      x-accepts: application/json
      x-tags:
      - tag: Documents
  /document/themes:
    get:
      operationId: getThemesOfDocument
      parameters:
      - description: The document id to identify the correct themes
        explode: true
        in: query
        name: documentId
        required: true
        schema:
          format: int64
          type: integer
        style: form
      responses:
        "200":
          content:
            '*/*':
              schema:
                items:
                  $ref: '#/components/schemas/RQMTheme'
                type: array
          description: Fetching themes successful
      summary: Get all themes of a document
      tags:
      - Themes
      x-accepts: '*/*'
      x-tags:
      - tag: Themes
  /document/links:
    get:
      operationId: getLinksOfDocument
      parameters:
      - description: The document id to identify the correct links
        explode: true
        in: query
        name: documentId
        required: true
        schema:
          format: int64
          type: integer
        style: form
      responses:
        "200":
          content:
            '*/*':
              schema:
                items:
                  $ref: '#/components/schemas/RQMLink'
                type: array
          description: Fetching links successful
      summary: Get all links of a document
      tags:
      - Documents
      x-accepts: '*/*'
      x-tags:
      - tag: Documents
...
  /elements:
    get:
      operationId: getElements
      parameters:
      - description: The document id for which the elements are fetched
        explode: true
        in: query
        name: documentId
        required: true
        schema:
          format: int64
          type: integer
        style: form
      responses:
        "200":
          content:
            '*/*':
              schema:
                items:
                  $ref: '#/components/schemas/RQMElement'
                type: array
          description: Elements successfully fetched
      summary: Get all elements
      tags:
      - Elements
      x-accepts: '*/*'
      x-tags:
      - tag: Elements
  /element:
    delete:
      operationId: deleteElement
      parameters:
      - description: The element to delete
        explode: true
        in: query
        name: elementId
        required: true
        schema:
          format: int64
          type: integer
        style: form
      responses:
        "200":
          description: Deleting element successful
      summary: Delete an element
      tags:
      - Elements
      x-accepts: application/json
      x-tags:
      - tag: Elements
    patch:
      operationId: patchElement
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RQMElement'
        description: The element to update
        required: true
      responses:
        "200":
          description: Updating element successful
      summary: Update an element
      tags:
      - Elements
      x-contentType: application/json
      x-accepts: application/json
      x-tags:
      - tag: Elements
    post:
      operationId: postElement
      parameters:
      - description: The rank of the element above, if no above element exists this
          shall be set to aaaaaaaaaaaaaaaaaaaa
        explode: true
        in: query
        name: aboveRank
        required: true
        schema:
          type: string
        style: form
      - description: The rank of the element below, if no below element exists this
          shall be set to an empty string
        explode: true
        in: query
        name: belowRank
        required: true
        schema:
          type: string
        style: form
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RQMElement'
        description: The element to create, if the element should be created at the
          highest level, the parentElementId shall be null
        required: true
      responses:
        "200":
          description: The element created by the backend
      summary: Add an element
      tags:
      - Elements
      x-contentType: application/json
      x-accepts: application/json
      x-tags:
      - tag: Elements
  /element/inlinks:
    get:
      operationId: getIncomingLinksOfElement
      parameters:
      - description: The id of the element
        explode: true
        in: query
        name: elementId
        required: true
        schema:
          format: int64
          type: integer
        style: form
      responses:
        "200":
          content:
            '*/*':
              schema:
                items:
                  $ref: '#/components/schemas/RQMLink'
                type: array
          description: Incoming links successfully fetched
      summary: Get all links of a given element
      tags:
      - Elements
      x-accepts: '*/*'
      x-tags:
      - tag: Elements
  /element/outlinks:
    get:
      operationId: getOutgoingLinksOfElement
      parameters:
      - description: The id of the element
        explode: true
        in: query
        name: elementId
        required: true
        schema:
          format: int64
          type: integer
        style: form
      responses:
        "200":
          content:
            '*/*':
              schema:
                items:
                  $ref: '#/components/schemas/RQMLink'
                type: array
          description: Outgoing links successfully fetched
      summary: Get all links of a given element
      tags:
      - Elements
      x-accepts: '*/*'
      x-tags:
      - tag: Elements
...
components:
...
  schemas:
...
    RQMDocument:
      example:
        approverId: 9
        confidentiality: confidentiality
        languageId: 7
        description: description
        previousBaselineId: 1
        authorId: 2
        baselineMinor: 4
        baselineMajor: 2
        lastModifiedOn: 2000-01-23T04:56:07.000+00:00
        baselineReview: 7
        internalIdentifier: 5
        lastModifiedById: 3
        reviewerText: reviewerText
        name: name
        externalIdentifier: externalIdentifier
        id: 1
        shortName: shortName
        workspaceId: 5
      properties:
        id:
          format: int64
          type: integer
        workspaceId:
          format: int64
          type: integer
        internalIdentifier:
          format: int64
          type: integer
        externalIdentifier:
          type: string
        name:
          type: string
        shortName:
          type: string
        description:
          type: string
        confidentiality:
          type: string
        authorId:
          format: int64
          type: integer
        languageId:
          format: int64
          type: integer
        approverId:
          format: int64
          type: integer
        reviewerText:
          type: string
        lastModifiedById:
          format: int64
          type: integer
        lastModifiedOn:
          format: date-time
          type: string
        baselineMajor:
          format: int64
          type: integer
        baselineMinor:
          format: int64
          type: integer
        baselineReview:
          format: int64
          type: integer
        previousBaselineId:
          format: int64
          type: integer
      type: object
    RQMElement:
      example:
        parentElementId: 5
        rank: rank
        documentId: 6
        elementTypeId: 1
        id: 0
        content: content
      properties:
        id:
          description: When creating an element the id is ignored by the backend because
            it generates one, when updating the element it has to be set
          format: int64
          type: integer
        documentId:
          description: When creating an element the documentId has to be set to assign
            the requirement to the correct document, it can not be updated.
          format: int64
          type: integer
        elementTypeId:
          format: int64
          type: integer
        content:
          type: string
        rank:
          type: string
        parentElementId:
          format: int64
          type: integer
      type: object
    RQMLink:
      example:
        fromDocumentShortName: fromDocumentShortName
        linkTypeId: 2
        fromElementId: 6
        toDocumentShortName: toDocumentShortName
        toDocumentId: 5
        id: 0
        fromDocumentId: 1
        toElementId: 5
      properties:
        id:
          description: Id of the link
          format: int64
          type: integer
        fromElementId:
          format: int64
          type: integer
        fromDocumentId:
          format: int64
          type: integer
        fromDocumentShortName:
          type: string
        toElementId:
          format: int64
          type: integer
        toDocumentId:
          format: int64
          type: integer
        toDocumentShortName:
          type: string
        linkTypeId:
          format: int64
          type: integer
      type: object
...
    RQMTheme:
      example:
        linkFromHighlightColor: linkFromHighlightColor
        linkToHighlightColor: linkToHighlightColor
        id: 0
      properties:
        id:
          format: int64
          type: integer
        linkFromHighlightColor:
          type: string
        linkToHighlightColor:
          type: string
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http